Acceptance/Rejection method change
Major: A/R now reuses the samples in the immediate previous iteration
(so the record of last_parameters is also no longer needed). Since the
samples after A/R at interation (n+1) is distributed exactly as
posterior in (n+1), these can be reused at (n+2), as compared to
previously the (n+2) iteration reuses samples from iteration n.

Minor changes:
1. Put generation of RV inside switch-cases to avoid redundant
generations when not necessary.
2. A bit more comments here and there.